msm: camera: isp: Update last applied scratch configuration for sfe
In case of sHDR dynamic switch, for a frame on which we are
configuring scratch buffer, use num of exposure from the
last applied valid request as opposed to global reference of
num of exposure held by global scratch configuration. The
global reference could have been updated by a future request,
it would be incorrect to use that when that request has not
even been applied.

CRs-Fixed: 3084946
Change-Id: I2a70d0de57e2dd780aa2eecc71a6398dc1800b6c
Signed-off-by: Depeng Shao <quic_depengs@quicinc.com>