qcacld-3.0: Add peer to inactive list based on unmap ref count
Currently dp peer is being added to inactive list during force
delete, but there is possiblity of peer getting freed up if
unmap ref count is not pending and leading to double free during
pdev cleanup.

To avoid this add the peer to inactive list only if unmap ref
count is pending.

Change-Id: Ib55a92dd36c4ec4a26d3103896ca31ff3d845148
CRs-Fixed: 3110728
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>