qcacld-3.0: Add bug_on if 5 consecutive ll_stats requests fails
If ll_stats command does not receive response from fw continuously
for some time then there is no way to determine the actual reason
of this comamand failure if the logs are overwritten.

With this change, add bug_on and trigger recovery after 5
continuous command failures so that the current logs are captured
to analyse the issue.

Change-Id: I8f2f08d598c25ccea5ecd4b02c86e1069daeb6e0
CRs-Fixed: 3106087
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>