qcacmn: Increase WMI CE buffer size to 3520 Bytes
As part of new change  to support large management frame
the buffer size of CE2 has been increased to 3520 Bytes.

Since CE2 & CE3 shares common EP in Fw because of which any changes
in buffer size in CE2 need to have similar changes on CE3 as well.
So to accommodate change increase CE3 buffer size.

Change-Id: I8e558f11b1c44934cf6d64d1f3027f016d34f526
CRs-Fixed: 3118928
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>