qcacld-3.0: Process deauth even in middle of SAE authentication
When SAE authentication fails due to some reason and if the
failure is detected while processing confirm request from the
station, some APs send deauth instead of confirm response.

Current driver/wpa_suplicant wait for the confirm response and
driver drops the deauth frame as it's in the middle of SAE
authentication. This results in connect timeout.
Driver is supposed to process the deauth from AP and indicate
failure to userspace. To resolve this, consider the deauth from AP
as SAE failure and abort the connection. Send connect failure to
userspace.

CRs-Fixed: 3157178
Change-Id: I09c9d8b19539e4986665c6052d3a7f52b3e49600
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>