qcacmn: Add check to skip connectivity logging
Currently we are sending DHCP and EAPOL connectivity
logs for Tx and Rx packets. but in case of DHCP
we are receiving DHCP DISC and DHCP REQ and
we are sending Tx status as TX_Fail.

So to fix issue ignore sending connectivity log in case
of Rx if packet type is DHCP DISC or DHCP REQ.

Change-Id: I688c375243cff0773bd8290e990d5106ff5b47ae
CRs-Fixed: 3164050
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>